requirements.txt
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.2
pandas_gbq==0.22.0
python-dotenv==1.0.1
scikit_learn==1.4.1.post1
scipy==1.13.1
seaborn==0.13.2
statsmodels==0.14.1
